refine the process for enabling/disabling dma
SPI DMA mode is disabled while update the touch firmware so the
DMA channel is not acquired. If the DMA mode is enabled again, it
needs to add a delay to make sure the SPI driver suspends then
the DMA channel will be acquired when the SPI driver resumes next
time.

Bug: 313710776
Test: No null pointer issue.

Change-Id: Iaf7f6ab869a461407aa85310efa26f7be90113d5
Signed-off-by: Yen-Chao Chen <davidycchen@google.com>