Move power state setting in the beginning of syna_dev_suspend
There is a corner case that when the syna_dev_suspend is running
and it triggers a reset which calls the helper work but screen
turns on at the same time. In the helper work we check the gti
pm_lock!=0 or power state!=PWR_ON to make sure the SPI commands
won't transfer while the touch driver suspends but in this corner
the power state is not change to PWR_OFF yet and pm_lock is set by
screen-on. Set the power state as PWR_OFF at the beginning to
avoid this issue.

Bug: 319765055
Bug: 318437757
Test: run suspned/resume test and check logs.

Change-Id: I86b4e4e6d83fcb178d1050aa1680341de425474c
Signed-off-by: Yen-Chao Chen <davidycchen@google.com>