[1/2] google-modules: edgetpu: Temporarily rollback to android-gs-caimito-6.1-android15
android-gs-caimito-6.1-android15
HEAD: 63bfb1d ("edgetpu: fix PAGE_SHIFT usages for non-4K support")

As of android15-qpr1-beta the GXP firmware ABI has been updated and
cross-compatibility with EdgeTPU's gcip-kernel-driver variant is
lost. Thus, we can no longer select a single gcip driver for both
modules to share.

This situation has proven to be extremely problematic, and more
time was spent trying to come up with a solution than I would have
preferred. No matter how you slice it, some abhorrent hacks will
ultimately be required to compile both versions of gcip while
avoiding collision.

Initially, the thought was to leverage objcopy to automate the
renaming of shared global functions. However, this came with the
undesirable side effect of breaking LTO.

As of right now, we can get away with rolling back to the 15
stable iterations of these modules in order to maintain gcip
cross-compatibility and make things a hell of a lot simpler in
the interim. This will buy some time to work out an alternative
approach that ticks all the boxes in a way that is satisfactory.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>